Use int for user & role ids

+++ Listopotamus.ApplicationCore/Entities/Identity/Role.cs
@@ -9,7 +9,7 @@ namespace Listopotamus.ApplicationCore.Entities.Identity
     /// <summary>
     /// Represents a role in the identity system.
     /// </summary>
-    public class Role : IdentityRole<long>
+    public class Role : IdentityRole<int>
     {
     }
 }
